+++ lib/action_policy/rspec/have_authorized_scope.rb
@@ -20,11 +20,12 @@ module ActionPolicy
     #   end
     #
     class HaveAuthorizedScope < ::RSpec::Matchers::BuiltIn::BaseMatcher
-      attr_reader :type, :name, :policy, :actual_scopes
+      attr_reader :type, :name, :policy, :scope_options, :actual_scopes
 
       def initialize(type)
         @type = type
         @name = :default
+        @scope_options = nil
       end
 
       def with(policy)
@@ -37,6 +38,11 @@ module ActionPolicy
         self
       end
 
+      def with_scope_options(scope_options)
+        @scope_options = scope_options
+        self
+      end
+
       def match(_expected, actual)
         raise "This matcher only supports block expectations" unless actual.is_a?(Proc)
 
@@ -44,7 +50,7 @@ module ActionPolicy
 
         @actual_scopes = ActionPolicy::Testing::AuthorizeTracker.scopings
 
-        actual_scopes.any? { |scope| scope.matches?(policy, type, name) }
+        actual_scopes.any? { |scope| scope.matches?(policy, type, name, scope_options) }
       end
 
       def does_not_match?(*)
@@ -57,8 +63,22 @@ module ActionPolicy
 
       def failure_message
         "expected a scoping named :#{name} for type :#{type} " \
+        "#{scope_options_message} " \
         "from #{policy} to have been applied, " \
         "but #{actual_scopes_message}"
+      end
+
+      def scope_options_message
+        if scope_options
+          if defined?(::RSpec::Matchers::Composable) &&
+             scope_options.is_a?(::RSpec::Matchers::Composable)
+            "with scope options #{scope_options.description}"
+          else
+            "with scope options #{scope_options}"
+          end
+        else
+          "without scope options"
+        end
       end
 
       def actual_scopes_message