+++ spec/action_policy/rspec_spec.rb
@@ -44,6 +44,10 @@ class TestService # :nodoc:
     authorized users, type: :data, with: CustomPolicy
   end
 
+  def filter_with_options(users, with_admins: false)
+    authorized users, type: :data, with: CustomPolicy, scope_options: { with_admins: with_admins }
+  end
+
   def own(users)
     authorized users, type: :data, as: :own, with: UserPolicy
   end
@@ -169,6 +173,18 @@ describe "ActionPolicy RSpec matchers" do
         expect { subject.own(target) }
           .to have_authorized_scope(:data).with(UserPolicy).as(:own)
       end
+
+      specify "with scope options" do
+        expect { subject.filter_with_options(target, with_admins: true) }
+          .to have_authorized_scope(:data).with(TestService::CustomPolicy)
+          .with_scope_options(with_admins: true)
+      end
+
+      specify "with composed scope options" do
+        expect { subject.filter_with_options(target, with_admins: true) }
+          .to have_authorized_scope(:data).with(TestService::CustomPolicy)
+          .with_scope_options(matching(with_admins: a_truthy_value))
+      end
     end
 
     context "when no scoping performed" do
@@ -178,7 +194,8 @@ describe "ActionPolicy RSpec matchers" do
             .to have_authorized_scope(:datum).with(TestService::CustomPolicy)
         end.to raise_error(
           RSpec::Expectations::ExpectationNotMetError,
-          %r{expected a scoping named :default for type :datum from TestService::CustomPolicy to have been applied}
+          Regexp.new("expected a scoping named :default for type :datum without scope options " \
+                     "from TestService::CustomPolicy to have been applied")
         )
       end
 
@@ -188,7 +205,8 @@ describe "ActionPolicy RSpec matchers" do
             .to have_authorized_scope(:data).with(UserPolicy)
         end.to raise_error(
           RSpec::Expectations::ExpectationNotMetError,
-          %r{expected a scoping named :default for type :data from UserPolicy to have been applied}
+          Regexp.new("expected a scoping named :default for type :data without scope options " \
+                     "from UserPolicy to have been applied")
         )
       end
 
@@ -198,7 +216,34 @@ describe "ActionPolicy RSpec matchers" do
             .to have_authorized_scope(:data).with(UserPolicy)
         end.to raise_error(
           RSpec::Expectations::ExpectationNotMetError,
-          %r{expected a scoping named :default for type :data from UserPolicy to have been applied}
+          Regexp.new("expected a scoping named :default for type :data without scope options " \
+                     "from UserPolicy to have been applied")
+        )
+      end
+
+      specify "scope options mismatch" do
+        expect do
+          expect { subject.filter_with_options(target, with_admins: true) }
+            .to have_authorized_scope(:data).with(TestService::CustomPolicy)
+            .with_scope_options(with_admins: false)
+        end.to raise_error(
+          RSpec::Expectations::ExpectationNotMetError,
+          Regexp.new("expected a scoping named :default for type :data " \
+                     "with scope options {:with_admins=>false} " \
+                     "from TestService::CustomPolicy to have been applied")
+        )
+      end
+
+      specify "composed scope options mismatch" do
+        expect do
+        expect { subject.filter_with_options(target, with_admins: true) }
+          .to have_authorized_scope(:data).with(TestService::CustomPolicy)
+          .with_scope_options(matching(with_admins: a_falsey_value))
+        end.to raise_error(
+          RSpec::Expectations::ExpectationNotMetError,
+          Regexp.new("expected a scoping named :default for type :data " \
+                     'with scope options matching {:with_admins=>\(a falsey value\)} ' \
+                     "from TestService::CustomPolicy to have been applied")
         )
       end
     end